Refactor clean code architecture and note UX
Centralizes DataStore enum observation, date formatting, shared note-list actions, editor navigation modes, and editor motion constants. Renames Room repository implementations with Impl suffixes, removes unused trash confirmation ids, expands KDoc for reusable Compose components, improves markdown checklist/list behavior, and updates Home/Agenda/Templates grids and floating action menu padding. Adds focused unit tests for DateUtils, editor modes, shared note actions, trash snackbar, and markdown formatting/parser behavior.

+++ app/src/main/java/io/github/r0x4nk/nexnote/ui/screen/agenda/AgendaActions.kt
@@ -22,7 +22,7 @@ internal data class AgendaActions(
     val onRequestTrash: (Note) -> Unit,
     val onRequestNoteActions: (Note) -> Unit,
     val onUndoTrash: (Long) -> Unit,
-    val onConfirmTrash: (Long) -> Unit
+    val onConfirmTrash: () -> Unit
 )
 
 @Composable
